requirements.txt
aniso8601==7.0.0
asgiref==3.3.1
astroid==2.5.1
boto3==1.19.11
botocore==1.22.11
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.6
dj-database-url==0.5.0
Django==3.1.7
django-cors-headers==3.7.0
django-filter==2.4.0
django-graphql-auth==0.3.16
django-graphql-jwt==0.3.0
django-model-mutations==0.1.1
django-storages==1.11.1
django-taggit==1.4.0
django-webpack-loader==1.1.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
django-phonenumber-field==5.2.0
Flask==1.1.2
google-api-core==1.30.0
google-auth==1.32.0
google-cloud-core==1.7.0
google-cloud-storage==1.39.0
google-crc32c==1.1.2
google-resumable-media==1.3.1
googleapis-common-protos==1.53.0
graphene==2.1.8
graphene-django==2.15.0
graphene-file-upload==1.3.0
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==20.1.0
idna==2.10
install==1.3.4
isort==5.7.0
itsdangerous==1.1.0
jmespath==0.10.0
lazy-object-proxy==1.5.2
packaging==20.9
Pillow==8.1.2
promise==2.3
protobuf==3.17.3
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.10.1
pydantic==1.8.2
PyJWT==1.7.0
pylint==2.7.2
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
requests==2.25.1
rsa==4.7.2
Rx==1.6.1
s3transfer==0.5.0
singledispatch==3.6.2
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
typing-extensions==3.10.0.0
urllib3==1.26.5
Werkzeug==1.0.1
whitenoise==5.2.0
wrapt==1.12.1
